+++ Runtime/AR/SubsystemLifecycleManager.cs
@@ -3,6 +3,10 @@ using System.Collections.Generic;
 
 using UnityEngine.XR.Management;
 
+#if UNITY_2020_2_OR_NEWER
+using UnityEngine.SubsystemsImplementation;
+#endif
+
 namespace UnityEngine.XR.ARFoundation
 {
     /// <summary>
@@ -10,9 +14,20 @@ namespace UnityEngine.XR.ARFoundation
     /// </summary>
     /// <typeparam name="TSubsystem">The <c>Subsystem</c> which provides this manager data.</typeparam>
     /// <typeparam name="TSubsystemDescriptor">The <c>SubsystemDescriptor</c> required to create the Subsystem.</typeparam>
-    public class SubsystemLifecycleManager<TSubsystem, TSubsystemDescriptor> : MonoBehaviour
+    public class SubsystemLifecycleManager<
+        TSubsystem, TSubsystemDescriptor
+#if UNITY_2020_2_OR_NEWER
+        , TProvider
+#endif
+        > : MonoBehaviour
+#if UNITY_2020_2_OR_NEWER
+        where TSubsystem : SubsystemWithProvider<TSubsystem, TSubsystemDescriptor, TProvider>, new()
+        where TSubsystemDescriptor : SubsystemDescriptorWithProvider<TSubsystem, TProvider>
+        where TProvider : SubsystemProvider<TSubsystem>
+#else
         where TSubsystem : Subsystem<TSubsystemDescriptor>
         where TSubsystemDescriptor : SubsystemDescriptor<TSubsystem>
+#endif
     {
         /// <summary>
         /// Get the <c>TSubsystem</c> whose lifetime this component manages.
@@ -25,7 +40,13 @@ namespace UnityEngine.XR.ARFoundation
         /// <value>
         /// The descriptor for the subsystem.
         /// </value>
-        public TSubsystemDescriptor descriptor => subsystem?.SubsystemDescriptor;
+        public TSubsystemDescriptor descriptor =>
+
+#if UNITY_2020_2_OR_NEWER
+            subsystem?.subsystemDescriptor;
+#else
+            subsystem?.SubsystemDescriptor;
+#endif
 
         /// <summary>
         /// Returns the active <c>TSubsystem</c> instance if present, otherwise returns null.